add PatchableFunctionEntry
the option: available as a string in SubstrateOptions, with a pair of
concealed options. the individual integers can probably be part of
GraalOptions, but the value checking happens in SubstrateOptions. we
might also need to check wherever we are using these things.

at compile time: All NOPs are added before emitting LIR to a
CompilationResult. The before/after is handled by adding an offset to
the function entry when the object file is laid out (this might cause an
issue around trampolines), and during patchMethods when fixing
rip-relative calls.

at runtime: the offset due to the nops is subtracted via
ImageCodeInfo.nopsBeforeEntry when the image info is loaded, so that
runtime relativeIP checks work out ok. the subtraction is done with
CodeInfoAccess, but not UntetheredCodeInfoAccess, which may cause issues
later.